requirements.txt
evdev
numpy
pyserial
pandas
matplotlib